BUGFIX: More savegame loading fixes
* Fix loading issues back to pre-1.0
* Be more robust about issues with files not being maps
* Avoid non-fatal error when there's no LastExportBonus

Fixes #540,#541
Also fixes a savegame reported in #bug-report
